fix: achieve PHPStan Level 9 compliance with empty baseline
- Remove redundant is_string() checks in CalculateQueueMetricsCommand
- Remove redundant ?? operators in RecordTrendDataCommand
- Add empty array guard before min()/max() in TrendAnalysisService
- Remove redundant isset() checks in PrometheusService
- Add proper type annotations and Redis method validation in LaravelQueueInspector
- Cast Carbon::now()->timestamp to int for binary operations
- Remove unused $driver variable in RedisBaselineRepository
- Clear PHPStan baseline from 32 errors to 0

+++ src/Repositories/RedisJobMetricsRepository.php
@@ -580,7 +580,7 @@ final readonly class RedisJobMetricsRepository implements JobMetricsRepository
             }
 
             $lastProcessedInt = is_numeric($lastProcessed) ? (int) $lastProcessed : 0;
-            $age = Carbon::now()->timestamp - $lastProcessedInt;
+            $age = (int) Carbon::now()->timestamp - $lastProcessedInt;
 
             if ($age > $olderThanSeconds) {
                 $driver->delete($key);
